team state, api calls, and routes hooked up
