+++ src/components/common/SecureRoute.js
@@ -1,4 +1,4 @@
-import React, { useEffect } from 'react';
+import React from 'react';
 import { SecureRoute } from '@okta/okta-react';
 import { Redirect, useLocation } from 'react-router-dom';
 import { connect } from 'react-redux';
