+++ src/components/pages/JoinTheSquad/RenderJoinTheSquad.js
@@ -1,24 +1,29 @@
-import React, { useEffect, useState } from 'react';
+import React, { useEffect } from 'react';
 import { Header } from '../../common';
 import { Col, Button } from 'antd';
-import { useHistory } from 'react-router-dom';
+// import { useHistory } from 'react-router-dom';
 import { useOktaAuth } from '@okta/okta-react/dist/OktaContext';
 import Squadup from '../../../assets/images/Squadup.svg';
 import wordbubble from '../../../assets/images/match_up_images/wordbubble.svg';
 import wordBubbleright from '../../../assets/images/match_up_images/wordBubbleright.svg';
-import {getChildSquad} from '../../../api';
+import {getChildTeam} from '../../../api';
+import { connect } from 'react-redux';
+
 
 const RenderJoinTheSquad = props => {
-  const { push } = useHistory();
+  // const { push } = useHistory();
   const {authState} = useOktaAuth();
 
   useEffect(() => {
-    getChildSquad(authState, props.child.id).then(
+    getChildTeam(authState, props.child.id).then(
       res => {
         console.log(res);
+        props.setTeamSubmissions(res);
       }
     );
-  });
+  // eslint-disable-next-line react-hooks/exhaustive-deps
+}, [authState]);
+
 
   return (
     <>
@@ -69,4 +74,9 @@ const RenderJoinTheSquad = props => {
   );
 };
 
-export default RenderJoinTheSquad;
+export default connect(
+  state => ({
+    team: state.team,
+  }),
+  {}
+)(RenderJoinTheSquad);
