+++ src/state/actions/index.js
@@ -2,6 +2,6 @@ import * as global from './globalActions';
 import * as child from './childActions';
 import * as parent from './parentActions';
 import * as tasks from './taskActions';
-import * as faceoffs from './faceoffsActions';
+import * as team from './teamActions';
 
-export { global, child, parent, tasks, faceoffs };
+export { global, child, parent, tasks, team };
