+++ src/state/reducers/index.js
@@ -5,11 +5,11 @@ import { combineReducers } from 'redux';
 import { reducer as child } from './childReducer';
 import { reducer as parent } from './parentReducer';
 import { reducer as tasks } from './taskReducer';
-import {reducer as faceoffs} from './faceoffsReducers';
+import { reducer as team } from './teamReducers';
 
 export default combineReducers({
   child,
   parent,
   tasks,
-  faceoffs,
+  team,
 });
